Move ignore related functions into language-js
