+++ src/language-css/printer-postcss.js
@@ -3,7 +3,6 @@
 const {
   printNumber,
   printString,
-  hasIgnoreComment,
   hasNewline,
   isFrontMatterNode,
   isNextLineEmpty,
@@ -1086,6 +1085,5 @@ module.exports = {
   print: genericPrint,
   embed,
   insertPragma,
-  hasPrettierIgnore: hasIgnoreComment,
   massageAstNode: clean,
 };
