+++ src/language-js/utils.js
@@ -5,8 +5,6 @@ const {
   getLast,
   hasNewline,
   hasNewlineInRange,
-  hasIgnoreComment,
-  hasNodeIgnoreComment,
   skipWhitespace,
 } = require("../common/util");
 const handleComments = require("./comments");
@@ -821,8 +819,8 @@ function hasJsxIgnoreComment(path) {
     prevSibling.type === "JSXExpressionContainer" &&
     prevSibling.expression.type === "JSXEmptyExpression" &&
     prevSibling.expression.comments &&
-    prevSibling.expression.comments.some(
-      (comment) => comment.value.trim() === "prettier-ignore"
+    prevSibling.expression.comments.some((comment) =>
+      isPrettierIgnoreComment(comment)
     )
   );
 }
@@ -1379,6 +1377,27 @@ function isBitwiseOperator(operator) {
   );
 }
 
+function isPrettierIgnoreComment(comment) {
+  return comment.value.trim() === "prettier-ignore";
+}
+
+function hasNodeIgnoreComment(node) {
+  return (
+    node &&
+    ((node.comments &&
+      node.comments.length > 0 &&
+      node.comments.some(
+        (comment) => isPrettierIgnoreComment(comment) && !comment.unignore
+      )) ||
+      node.prettierIgnore)
+  );
+}
+
+function hasIgnoreComment(path) {
+  const node = path.getValue();
+  return hasNodeIgnoreComment(node);
+}
+
 module.exports = {
   classChildNeedsASIProtection,
   classPropMayCauseASIProblems,
@@ -1399,6 +1418,8 @@ module.exports = {
   hasSameLoc,
   hasTrailingComment,
   hasTrailingLineComment,
+  hasIgnoreComment,
+  hasNodeIgnoreComment,
   identity,
   isBinaryish,
   isCallOrOptionalCallExpression,
